STEP3 - 주문에 따른 빼빼로 그리기
